Filter events by the past as well as the future
In order to do this, I also improved the interface for the filters to
rely on a custom `Filter` type instead of a `Maybe` for more expressive
filtering.

I also wrapped the event list we expose to other parts of the app so it
doesn't need to worry about the filtering and just provides the events.