Remove partner logo from show partner page
This data comes from PlaceCal where partner admins can upload any
arbitrary image (not just a logo) so sometimes this is a picture
of their group or facilities. Furthermore: if this is a logo we
don't have any way of making the logo look correct on the TD page
which is important if they have a transparent PNG with dark text.